add junit coverage to the AssignTask flow
